pyproject.toml
[tool.poetry]
name = "fsreading-backend"
version = "0.1.0"
description = ""
authors = ["fengshu <bfstudy@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
ruff = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
apollo-client-py = "^0.1.1"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
